requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
azure-ai-textanalytics==5.3.0
azure-cognitiveservices-vision-computervision==0.9.1
azure-common==1.1.28
azure-core==1.32.0
azure-search-documents==11.5.2
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
distro==1.9.0
Flask==3.1.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
idna==3.10
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.8.0
MarkupSafe==3.0.2
msrest==0.7.1
oauthlib==3.2.2
openai==1.55.3
outcome==1.3.0.post0
pillow==11.0.0
pydantic==2.10.2
pydantic_core==2.27.1
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==2.0.0
selenium==4.27.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
tqdm==4.67.1
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.3
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
